requirements.txt
# AI used to generate this comment below:
# List of Python packages required for this project
# Tested with Python 3.13.1
# Use a virtual environment to avoid conflicts
# Install with: pip install -r requirements.txt
anyio==4.11.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.2
firebase_admin==7.1.0
gcloud==0.18.3
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jwcrypto==1.5.6
msgpack==1.1.2
oauth2client==4.1.3
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.2.5
Pyrebase4==4.8.0
python-dotenv==1.1.1
python-jwt==4.1.0
requests==2.32.5
requests-toolbelt==0.10.1
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==1.26.20
wheel==0.45.1
